+++ pkg/aws/g8scontrolplane/mutate_g8scontrolplane_infraref_test.go
@@ -131,6 +131,7 @@ func getG8sControlPlaneNoRefRAWByte() ([]byte, error) {
 			Namespace: controlPlaneNameSpace,
 			Labels: map[string]string{
 				"giantswarm.io/control-plane":            controlPlaneName,
+				"giantswarm.io/cluster":                  clusterName,
 				"cluster-operator.giantswarm.io/version": "1.2.3",
 				"giantswarm.io/organization":             "giantswarm",
 				"release.giantswarm.io/version":          "11.5.0",